requirements.txt
cryptography
hexdump
pycryptodomex
pyscard
websockets
